cmd/atlasci/internal/ci: add basic run structure
The general idea is to wrap these analyzers with top-level
runner that can be configured with custom reporter/writer
and setup driver-specific configuration.

On the execution part, the runner should be responsible for
scanning (and parsing, soon) the SQL files in the migration
directory, create a pass (check context) for each new one and
execute the registered and desired analyzers on it.

The reporting should be environment-aware. e.g. GitHub PR,
local development (HCL, Ent, plain SQL, etc).