sql/migrate: change execution logic (#939)
* sql/migrate: change execution logic

Instead of having a state for a migration execution like ok, error, ... we now have a counter or every executed statement. This will increase for every successfully applied statement. Every next attempt to migrate, will pick up the file where it left off and attempt to execute the new statement.