cmd/atlas/internal/cmdapi: schema diff works with statereaders
This is the first PR in a series of PRs to enable all schema and migrate commands to work with `migrate.StateReader` to enable the CLI to read and diff states between every supported state definitions (currently HCL, migration directory and database connection).

As a second goal the PR's aim to increase the amount of code shared between `schema` and  `migrate` commands and refactor the code structure to increase readability and reduce global state (which is incredibly painful when unit testing).