sql/migrate: ensure consistency of NotCleanError
Some functions were returning migrate.NotCleanError value, some other - pointer.
This was the cause of inconsistent error matching behavior.

It is dangerous to use a non-pointer to implement the error interface.
Following example probably should give more clues about the issue:
  var cerr *NotCleanError
  errors.As(NotCleanError{}, &cerr) // false
  errors.As(&NotCleanError{}, &cerr) // true

With this change it is now impossible to return NotCleanError value, but
pointer. This also means that stuff like errors.As(err,
&migrate.NotCleanError{}) isn't going to work anymore, which is not a bad thing
because errors.As's target argument is supposed be a pointer to a value of the
error type:
  var cerr *NotCleanError
  errors.As(err, &cerr)

See https://pkg.go.dev/errors#As and https://go.dev/blog/go1.13-errors for more
info.