cmd/atlas/internal/migratelint: make linting inspection context-aware
If the connection is scoped to a specific schema, we should use InspectSchema to
get the inspected realm. That is because PostgreSQL unsets the search_path when
inspecting the entire realm, causing us to lose the schema context and get qualified
composite types on inspection.

Fixed https://github.com/ariga/atlas/issues/3200